Improve screenshot quality
We add support for specifying hostname-specific sleep intervals and
vertical screenshot offsets. This is necessary because some pages are
just too different, e.g. placing generic text or jobsearch
infrastructure at the top of the job ad.

Rules are ranked according to specificity, e.g. a rule defined for
subdoman.domain.com takes precedence over a rule defined for
domain.com (but domain.com is still taken into account).

The rules are defined in a JSON file pointed to by `imageConfPath`.
If that variable contains a URL string, the config file is fetched
via HTTP - otherwise it is read from the local filesystem.

Additionally, we replace event-listening with a standard sleep delay.
Listening for certain browser events is a brittle strategy when the
intended targets are (almost) arbitrary websites; this approach seems
to produce better results.