Make it clear when code is lifted from other copyright holders
Discussed with / approved by KSM/ksmadsen.

Most of us would agree that code reuse is a good thing! However, not
all pieces of reusable code are wrapped up nicely as external
libraries, and even when they are, the size and added complexity of
the import might not be worth it. Conversely, if we just copy-paste
bits of code here and there, that leaves open the question of proper
attribution.

Since Spectura is still a small project, and to keep things simple,
we'll continue to assume that typical contributions are done by a
single copyright holder (specified in the LICENSE file). For
situations where external code is copied into the repository, we'll
use the xlib folder and separate copyright headers going forward.

This commit does exactly that, while replacing all byte size
formatting with an implementation from beyond. The new function
supports different precisions, while consistently handling powers
of two (i.e. 1000 * 1024 is formatted as "1000 KB", not "1 MB").