Auto-refresh images in the background (#7)
We want Spectura to auto-refresh cached screenshots once in a while.
There are two reasons for this:

* Broken screenshots might be caused by ephemeral issues (e.g.
  network problems, slow web servers, heavy load on Decap), and
  simply trying again might fix the everything.

* The content of the web site might change, outdating the
  screenshot. We want the system to be somewhat self-healing, to
  avoid handling too many out-of-sync issues manually.

To avoid replacing a good image with a bad one, the refreshed image
must have higher information density than the earlier version. We
define information density by the absense of large areas with
monochrome pixels. (See the calculateScore function for details).